pkg/video/ffmpeg: to help with indexing videos
Plus all the changes needed (indexer, search, server) to make use of it
in the web UI, and in particular to generate and display an image
thumbnail for a video.

There are 2 potentially controversial changes in particular that might
require more careful thinking:

1)
The video attributes are keyed by wholeRef, not by fileRef. Same in
corpus. This is in opposition with the image attributes, but in line
with gps coordinates, media (music) attributes, and EXIF attributes.

2)
Because thumbnail.Service is an exported function, I think it would be
gross to pass it a jsonconfig.Obj as a parameter, and passing a
serverconfig.VideoThumbnail seems like a nicer API (not to mention it's
a documented type). So when generating the low-level config, we could
convert the high-level serverconfig.VideoThumbnail	to a
jsonconfig.Obj, and reconvert it to a serverconfig.VideoThumbnail
before passing it to thumbnail.Service (so probably in
pkg/server/ui.go). Or, we could simply encode/decode as JSON, which is
both simpler and what people are used to, and which is what I chose to
do instead.

Also I haven't written any index upgrade path so far, as it's not a
critical index change, and I'm waiting for the rest of the CL to be OKed
first anyway.

Credits to Antonin Amand for the initial work on this.

Fixes #791

Change-Id: Ica5dbc30b5552dff5b1cf6af20fcab209ff81a5f